server/pyproject.toml
[tool.poetry]
name = "q4-capstone"
version = "0.1.0"
description = ""
authors = ["Jacob-Short <thatmanjshort@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
Pillow = "^8.3.2"
requests = "^2.26.0"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
